feat(libcodegen): add generate/download subcommands, gate proto toolchain (#2220)
Split fit-codegen into a thin dispatcher plus src/commands/generate.js (today's
generation body, with the proto-compiler toolchain loaded via guarded dynamic
imports) and src/commands/download.js (folds in fit-download-bundle's fetch).
Move @grpc/proto-loader, mustache, and protobufjs-cli to optionalDependencies so
a download-only production install omits them; generate prints a reinstall hint
when they are absent. Add libconfig as a direct dependency for download's
createScriptConfig. Bare fit-codegen now shows help (exit 0), the subcommand
front door; the no-flags usage error moves under generate. execLine shifts one
argv slot to skip the download subcommand token.
