+++ libraries/libcodegen/package.json
@@ -46,13 +46,16 @@
   },
   "dependencies": {
     "@forwardimpact/libcli": "^0.1.0",
+    "@forwardimpact/libconfig": "^0.1.79",
     "@forwardimpact/libpreflight": "^0.1.0",
     "@forwardimpact/libstorage": "^0.1.53",
     "@forwardimpact/libtelemetry": "^0.1.22",
     "@forwardimpact/libutil": "^0.1.64",
+    "protobufjs": "^7.6.3"
+  },
+  "optionalDependencies": {
     "@grpc/proto-loader": "^0.8.1",
     "mustache": "^4.2.0",
-    "protobufjs": "^7.6.3",
     "protobufjs-cli": "1.3.3"
   },
   "devDependencies": {
